+++ packages/@vangware/types/src/updater/generateJSDoc.ts
@@ -1,0 +1,38 @@
+import type { ITagData } from "vscode-html-languageservice/lib/esm/htmlLanguageTypes.js";
+import type { ReadOnly } from "../ReadOnly.js";
+import { getDescription } from "./getDescription.js";
+import { normalizeJSDocMarkdown } from "./normalizeJSDocMarkdown.js";
+
+/**
+ * Generate formatted JSDoc.
+ *
+ * @category Internal
+ * @param options Description and references to be formatted.
+ * @returns Formatted JSDoc.
+ */
+export const generateJSDoc = ({
+	description,
+	references,
+}: {
+	readonly description?: string | { readonly value: string };
+	readonly references?: ReadOnly<ITagData["references"]>;
+}) =>
+	description === undefined
+		? "\n"
+		: `
+/**
+ * ${normalizeJSDocMarkdown(getDescription(description))}${
+		references
+			? `
+ * 
+ * ---
+ *
+ * **References**
+ * 
+ * ${references
+		.map(reference => `@see [${reference.name}](${reference.url})`)
+		.join("\n * ")}`
+			: ""
+ }
+ */
+`;