+++ packages/@vangware/types/src/updater/normalizeJSDocMarkdown.ts
@@ -1,0 +1,12 @@
+/**
+ * Markdown can have new lines and lists, this normalizes everything.
+ *
+ * @category Internal
+ * @param markdown Markdown to normalize.
+ * @returns Normalized markdown.
+ */
+export const normalizeJSDocMarkdown = (markdown: string) =>
+	markdown
+		.replaceAll("\n", "\n * ")
+		.replaceAll("\n **   ", "\n * -   ")
+		.replaceAll(" ", " ");