+++ packages/@vangware/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## 6.3.0
+
+### Minor Changes
+
+-   ✨ add new HTML attributes types.
+
 ## 6.2.10
 
 ### Patch Changes